co_code 타입 수정, evaluation 관련 jsp 수정, co_code 세션에서 일부 못 가져온 부분 수정

+++ khis/src/main/java/com/kh/khis/interview_evaluation/model/vo/Apply.java
@@ -18,7 +18,7 @@ public class Apply implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	private int member_info_no;
-	private int co_code;
+	private long co_code;
 	@JsonFormat(shape=JsonFormat.Shape.STRING, pattern="yyyy-MM-dd")
 	private Date start_time;
 	@JsonFormat(shape=JsonFormat.Shape.STRING, pattern="yyyy-MM-dd")
